backend/auth_service/requirements.txt
fastapi
uvicorn
pyjwt
argon2-cffi
pyodbc
requests
